ruse/compiler: Generalise constant binary op parameter handling
Previously, we had special handling for the two binary shift
operations when their second parameter is a constant, as in
(<< foo 3).

This commit generalises that to any binary operation that supports
constant parameters. It's just the compiler stage for now, but it
was enough change to have in its own commit.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>