ruse/compiler: Added support for constant args to more operators
This extends the support for constant operands to more operations.
Previously, we supported constant shifts left (<<) and right (>>).
This commit adds support on x86-64 to add (+), sub (-), or (or),
and (and), xor (xor), equal (=), not equal (!=), less than (<),
less than or equal (<=), greater than (>), greater than or equal
(>=).

We also add some tests.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>