ruse: Finish support for expression comments
We already had basic support for expression comments, where an
entire expression is commented out in a single go, by prefixing
it with a hash and a semicolon (example: #;(complex (list))).
This initial support consisted of parsing expression comments
and discarding them.

This commit adds further support in three stages:
- There is now an AST node to represent an expression comment.
- The expressions in a file are now ast.Expression, not *ast.List.
- The parser will emit expression comment nodes if Mode sets ParseComments.
- The type checker and compiler will ignore expression comments.
- The formatter correctly prints expression comments.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>