ruse: Embed the standard library to simplify deployments
This is a little confusing but not too complex. We rename the old
Ruse tool target from :ruse to :ruse_nostdlib and use that as the
tool in ruse_internal_library and ruse_standard_library. We then
build a second version of the Ruse tool with an extra source file,
which embeds the standard library (which is now :ruse).

This means that we can now use the standard library built with the
Ruse tool without needing to keep a copy on disk.

Signed-off-by: SlyMarbo <the.sly.marbo@googlemail.com>